--- projects/constantin/src/linalg/VoxelGrid.ts
@@ -1,16 +1,0 @@
-
-
-
-class VoxeGrid{
-
-
-    calculateIndicesForPosition(x:number,y:number,z:number){
-
-    }
-
-    calculateIndicesForPosition2(x:number,y:number){
-        let ret=new Array<number>();
-        let idx0
-
-    }
-}
